fix: s/usize/f64/

+++ content/blog/2025-01-01-thin_delegate_aux/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod enum_dispatch {
     pub struct Rect {
-        width: usize,
-        height: usize,
+        width: f64,
+        height: f64,
     }
 
     #[enum_dispatch::enum_dispatch]
@@ -11,11 +11,11 @@ pub mod enum_dispatch {
 
     #[enum_dispatch::enum_dispatch(Shape)]
     pub trait ShapeI {
-        fn area(&self) -> usize;
+        fn area(&self) -> f64;
     }
 
     impl ShapeI for Rect {
-        fn area(&self) -> usize {
+        fn area(&self) -> f64 {
             self.width * self.height
         }
     }
