+++ content/blog/2025-01-01-thin_delegate_aux/tests/ui/enum_delegate_v030/fail_not_supported_associated_const.rs
@@ -15,12 +15,12 @@ enum Shape {
 }
 
 struct Rect {
-    width: usize,
-    height: usize,
+    width: f64,
+    height: f64,
 }
 
 impl ShapeI for Rect {
-    type Output = usize;
+    type Output = f64;
     const NAME: &'static str = "Rect";
 
     fn area(&self) -> Self::Output {
