+++ content/blog/2025-01-01-thin_delegate_aux/tests/ui/enum_dispatch/fail_implicit_error_if_typo.rs
@@ -1,11 +1,11 @@
 #[enum_dispatch::enum_dispatch]
 trait ShapeI {
-    fn area(&self) -> usize;
+    fn area(&self) -> f64;
 }
 
 struct Rect {
-    width: usize,
-    height: usize,
+    width: f64,
+    height: f64,
 }
 
 // enum_dispatch can't detect typos.
@@ -15,14 +15,14 @@ enum Shape {
 }
 
 impl ShapeI for Rect {
-    fn area(&self) -> usize {
+    fn area(&self) -> f64 {
         self.width * self.height
     }
 }
 
 fn main() {
-    let rect = Rect { width: 2, height: 3 };
-    assert_eq!(rect.area(), 6);
+    let rect = Rect { width: 2.0, height: 3.0 };
+    assert_eq!(rect.area(), 6.0);
     let shape = Shape::Rect(rect);
-    assert_eq!(shape.area(), 6);
+    assert_eq!(shape.area(), 6.0);
 }
